Rescue any startup execptions and retry startup
This change addresses failure cases during startup. While it was
designed with dns resolution failures causing the client to crash in
mind, there may be any number of exceptions that get thrown from
places like eventmachine, etc. that we would want to handle gracefully.

First we catch any exceptions and log the error and then back off
restarting the client for [:client][:restart_wait] seconds. (This
defaults to 10).  We do this indefinitely until the client has startup.