Fix occurrences count & reference original event.
Change handling of event occurrences so that the count is only reset on
an OK event: If an event is flapping between states and no flapping
thresholds have been set its possible for it to never pass the
occurrences threshold for the event to be handled. This will now
continually increment occurrences so long as the check is reporting
events without an OK status.

Additionally, added :first_id to the event data which will always
contain the :id of the first event in this series of events, and kept so
long as the events are not reporting OK or check is in a flapping state.
This can be useful to help reference a batch of events, for example by
using the :id and :first_id in the Message-ID or In-Reply-To headers of
e-mails.